requirements.txt
autogen-agentchat
autogen-ext[openai]
playwright
python-dotenv
requests
Pillow
qrcode[pil]
google-genai